Create README.md
As part of the mobile design module for university, I was tasked with creating an android application in which users can create travel journals for holidays. The user creates a holiday, and stores posts and locations that they visited during that holiday.

In addition, the user can find places of interest nearby and use an interactive map to see the locations of previous posts.